requirements.txt
pactus-sdk>=1.2.4
eth-account>=0.13.7
python-dotenv>=1.0.0
web3>=6.0.0
4